Move modals to single location
