+++ apps/client/src/pages/sales/journeys/components/index.ts
@@ -1,3 +1,3 @@
-export { DeleteJourneyModal } from './DeleteJourneyModal';
-export { TrackJourneyModal } from './TrackJourneyModal';
-export { UntrackJourneyModal } from './UntrackJourneyModal';
+export { DeleteJourneyModal } from '@/components/modals/delete-journey-modal';
+export { TrackJourneyModal } from '@/components/modals/track-journey-modal';
+export { UntrackJourneyModal } from '@/components/modals/untrack-journey-modal';
